services/kafka-consumers/influx-writer/requirements.txt
influxdb-client>=1.48.0
kafka-python>=2.0.2
